replace tones list with degrees dict
tones was a list of tones in cents above the root. degrees is a dict of degree->tone, so list(degrees.values()) is equivalent to the old tones list.
Did this to avoid potential confusion in the future with tones being 0-based list (would always have to add 1 to the index to get the degree).
Perhaps there is a better struct for this (pandas Series)?